image-loader/requirements.txt
fastapi==0.88.0
numpy==1.24.0
Pillow==9.3.0
torch==1.13.1
torchvision==0.14.1
uvicorn==0.20.0
